+++ client/facetome/src/components/CreatePost.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef } from "react";
-import { createPost } from "../apiServices/posts";
+import { createPost } from "../apiRequests/posts";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircleNotch } from "@fortawesome/free-solid-svg-icons";
 
@@ -67,17 +67,17 @@ const CreatePost = ({
     }
   }, [receiveData]);
 
-  const autoResize = () => {// resize textarea based on scrollHeight
-  if (contentRef.current) {
-    contentRef.current.style.height = 'auto'; 
-    contentRef.current.style.height = contentRef.current.scrollHeight + 'px'; 
-  }
-};
+  const autoResize = () => {
+    // resize textarea based on scrollHeight
+    if (contentRef.current) {
+      contentRef.current.style.height = "auto";
+      contentRef.current.style.height = contentRef.current.scrollHeight + "px";
+    }
+  };
 
-useEffect(() => {
-  autoResize();
-}, [content]);
-
+  useEffect(() => {
+    autoResize();
+  }, [content]);
 
   return (
     <div className="bg-2 p-4  text-white rounded-xl w-full shadow-xl">
@@ -93,26 +93,25 @@ useEffect(() => {
           required
           className="p-2 my-2 rounded-xl w-full text-4 bg-gray-100"
         />
-       <div className="flex flex-col sm:flex-row sm:justify-between sm:items-center gap-2 sm:gap-4">
-  <label htmlFor="content" className="text-l font-bold pl-1 sm:pl-3">
-    Content
-  </label>
-  {setStep && (
-    <div className="text-sm font-medium text-white italic text-right sm:text-left">
-      Review and edit the generated post.
-    </div>
-  )}
-</div>
-
+        <div className="flex flex-col sm:flex-row sm:justify-between sm:items-center gap-2 sm:gap-4">
+          <label htmlFor="content" className="text-l font-bold pl-1 sm:pl-3">
+            Content
+          </label>
+          {setStep && (
+            <div className="text-sm font-medium text-white italic text-right sm:text-left">
+              Review and edit the generated post.
+            </div>
+          )}
+        </div>
 
         <textarea
-         ref={contentRef} 
+          ref={contentRef}
           id="content"
           value={content}
           onChange={(e) => {
-    setContent(e.target.value);
-    autoResize();
-  }}
+            setContent(e.target.value);
+            autoResize();
+          }}
           rows="2"
           required
           className="p-2  rounded-xl w-full text-4 bg-gray-100"
