+++ client/facetome/src/pages/FindFriendsPage.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { getAllUsers } from "../apiServices/users";
+import { getAllUsers } from "../apiRequests/users";
 import NavBar from "../components/NavBar";
 import FollowButton from "../components/FollowButton";
 import FollowRequestResponse from "../components/FollowRequestResponse"; // Adjust path as needed
@@ -63,7 +63,7 @@ const FindFriendsPage = () => {
                 status = "Pending";
               } else if (user.hasCTA) {
                 status = "Requested";
-              }else if (user.canFollow) {
+              } else if (user.canFollow) {
                 status = "";
               }
